third_party/WebKit/LayoutTests/css3/device-adapt/opera/constrain-007-expected.txt


FAIL CSS Test: @viewport constrained - width 50% assert_equals: expected 160 but got 400
